requirements.txt
keras~=2.15.0
numpy~=1.26.2
tensorflow~=2.15.0.post1
scikit-learn~=1.4.1.post1
enlighten~=1.12.4
matplotlib~=3.8.2
torch~=2.2.1
Pillow~=10.1.0
alive-progress~=3.1.5
imageio~=2.33.1